Move bronze mappings into harvest/mappings.
Group the OpenSearch bronze mappings in one package and point the index command and Painless comments at the new paths.
