+++ harvest/mappings/bronze_article.py
