+++ harvest/mappings/bronze_social_production.py
